Fix the logic properly and move the template to header
