+++ hardware_interface/include/hardware_interface/lexical_casts.hpp
@@ -15,7 +15,11 @@
 #ifndef HARDWARE_INTERFACE__LEXICAL_CASTS_HPP_
 #define HARDWARE_INTERFACE__LEXICAL_CASTS_HPP_
 
+#include <regex>
+#include <sstream>
+#include <stdexcept>
 #include <string>
+#include <type_traits>
 #include <vector>
 
 namespace hardware_interface
@@ -29,6 +33,77 @@ namespace hardware_interface
 double stod(const std::string & s);
 
 bool parse_bool(const std::string & bool_string);
+
+template <typename T>
+std::vector<T> parse_array(const std::string & array_string)
+{
+  // Use regex to check for a flat array: starts with [, ends with ], no nested brackets
+  const std::regex array_regex(R"(^\[\s*([^\[\]]*\s*(,\s*[^\[\]]+\s*)*)?\]$)");
+  if (!std::regex_match(array_string, array_regex))
+  {
+    throw std::invalid_argument(
+      "String must be a flat array: starts with '[' and ends with ']', no nested arrays");
+  }
+
+  // Use regex for the expression that either empty or contains only spaces
+  const std::regex empty_or_spaces_regex(R"(^\[\s*\]$)");
+  if (std::regex_match(array_string, empty_or_spaces_regex))
+  {
+    std::cerr << "Input is an empty array: " << array_string << std::endl;
+    return {};  // Return empty array if input is "[]"
+  }
+
+  // Use regex to find cases of comma-separated but only whitespaces or no spaces between them like
+  // "[,]" "[a,b,,c]"
+  const std::regex comma_separated_regex(R"(^\[\s*([^,\s]+(\s*,\s*[^,\s]+)*)?\s*\]$)");
+  if (!std::regex_match(array_string, comma_separated_regex))
+  {
+    throw std::invalid_argument(
+      "String must be a flat array with comma-separated values and no spaces between them");
+  }
+
+  std::vector<T> result = {};
+  if (array_string == "[]")
+  {
+    return result;  // Return empty array if input is "[]"
+  }
+
+  // regex for comma separated values and no spaces between them or just content like "[a,b,c]" or
+  // "[a]" or "[a, b, c]"
+  const std::regex value_regex(R"([^\s,\[\]]+)");
+  auto begin = std::sregex_iterator(array_string.begin(), array_string.end(), value_regex);
+  auto end = std::sregex_iterator();
+
+  for (auto it = begin; it != end; ++it)
+  {
+    const std::string value_str = it->str();  // Get the first capturing group
+    if constexpr (std::is_same_v<T, std::string>)
+    {
+      result.push_back(value_str);
+    }
+    else if constexpr (std::is_same_v<T, bool>)
+    {
+      result.push_back(parse_bool(value_str));
+    }
+    else if constexpr (std::is_floating_point_v<T> || std::is_integral_v<T>)
+    {
+      if (const auto value = hardware_interface::stod(value_str))
+      {
+        result.push_back(static_cast<T>(value));
+      }
+      else
+      {
+        throw std::invalid_argument(
+          "Failed converting string to floating point or integer: " + value_str);
+      }
+    }
+    else
+    {
+      throw std::invalid_argument("Unsupported type for parsing: " + std::string(typeid(T).name()));
+    }
+  }
+  return result;
+}
 
 std::vector<std::string> parse_string_array(const std::string & string_array_string);
